refactor: moved userActions to own file
